+++ cuppa/toolchains/gcc.py
@@ -608,6 +608,18 @@ class Gcc(object):
         return [ '-fmodules', '-fmodule-mapper={}'.format( mapper_path( env ) ) ]
 
 
+    def profiles_supported( self, env ):
+        return False
+
+
+    def profiles_enable_flags( self, env ):
+        return []
+
+
+    def profiles_enforce_flags( self, env, names ):
+        return []
+
+
     def module_bmi_path( self, env, module_name ):
         from cuppa.toolchains.cxx_modules_support import named_bmi_path
         return named_bmi_path( env, module_name, '.gcm' )